+++ test/orchard/inspect_test.clj
@@ -587,7 +587,7 @@
                   (inspect/down 40)
                   (inspect/next-sibling)
                   :value))))
-  (testing "next sibling doesn't go beyond the current page"
+  (testing "next sibling does go beyond the current page"
     (is (= 3 (-> (list 1 2 3)
                  inspect
                  (inspect/down 2)
@@ -597,10 +597,23 @@
                  (inspect/next-sibling)
                  (inspect/next-sibling)
                  :value)))
-    (is (= 41 (-> long-vector
+    (is (= 45 (-> long-vector
                   inspect
-                  (inspect/set-page-size 6)
+                  (inspect/set-page-size 3)
                   (inspect/down 40)
+                  (inspect/next-sibling)
+                  (inspect/next-sibling)
+                  (inspect/next-sibling)
+                  (inspect/next-sibling)
+                  (inspect/next-sibling)
+                  (inspect/next-sibling)
+                  :value))))
+  (testing "next-sibling doesn't fall beyond the last element."
+    (is (= 69 (-> long-vector
+                  inspect
+                  (inspect/set-page-size 1)
+                  (inspect/down 67)
+                  (inspect/next-sibling)
                   (inspect/next-sibling)
                   (inspect/next-sibling)
                   (inspect/next-sibling)